2.0 Algoritmos funcionais
Pelos primeiros testes que fiz, os algoritmos estão funcionando bem,
exceto para criação de chaves usando os números primos 2,3.

Para criar as chaves, use a seguinte sintaxe:
[e, d, n ] = RSA_keys(num_primo1, num_primo2). Ele vai gravar os
atributos "e", "d" e "n" em suas respectivas variáveis, assim como é
descrito no PDF. OBS: Se quiserem, rodem "testMiller.m" para obter uma
lista de primos até 1000.

Para encriptografar algo, use a seguinte sintaxe:
encriptografado = rsa_cript(texto_plano,e,n)
onde "texto_plano" é um inteiro qualquer que você queira encriptografar.

Para desencriptografar use a seguinte sintaxe:
texto_plano = rsa_decript(encriptografado, d,n)

Lembrando que [e,n] e [d,n] são, respectivamente, as chaves pública e
privada do RSA.

+++ exponenciacaoMod.m
@@ -2,7 +2,7 @@ function [ resp ] = exponenciacaoMod( a, b, n )
 
 c = 0;
 f = 1;
-k = floor(log2(b));
+k = floor(log2(double(b)));
 for i = k:-1:0
     c = 2*c;
     f = mod(f*f,n);
